requirements.txt
alembic==1.13.2
anyio==4.4.0
APScheduler==3.10.4
arrow==1.3.0
attrs==23.2.0
backports.functools-lru-cache==1.3
beautifulsoup4==4.12.3
black==24.4.2
certifi==2024.7.4
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
dateparser==1.2.0
deep-translator==1.11.4
distlib==0.3.8
filelock==3.15.4
fqdn==1.5.1
greenlet==3.0.3
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.8.1
httpcore==1.0.5
httpx==0.27.0
hyperframe==5.2.0
idna==2.10
isoduration==20.11.0
jsonpointer==3.0.0
lxml==5.2.2
Mako==1.3.5
MarkupSafe==2.1.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
outcome==1.3.0.post0
packaging==24.1
pathspec==0.12.1
pip-autoremove==0.10.0
pipdeptree==2.23.1
pipenv==2024.0.1
platformdirs==4.2.2
postal==1.1.10
pullenti-client==0.6.0
pymorphy2-dicts-ru==2.4.417127.4579844
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.4
pytz==2024.1
redis==5.0.8
regex==2024.7.24
requests==2.32.3
rfc3986==1.5.0
rq==1.16.2
selenium==4.23.1
setuptools==72.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.31
trio==0.26.0
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzlocal==5.2
uri-template==1.3.0
urllib3==2.2.2
virtualenv==20.26.3
webcolors==24.6.0
websocket-client==1.8.0
wsproto==1.2.0
